server/requirements.txt
fastapi
uvicorn[standard]
numpy
scipy
plyfile
msplat[cli]
torch
Pillow
realesrgan
basicsr
